backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ffmpeg-python==0.2.0
watchdog==3.0.0
send2trash==1.8.2
pathvalidate==3.2.0
python-magic==0.4.27
structlog==23.2.0
python-multipart==0.0.6
cors==1.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
psutil==5.9.6
pillow==10.1.0
firebase-admin==6.4.0